move testing_methods to methods directory
