+++ pysat/instruments/methods/__init__.py
@@ -5,4 +5,4 @@ specific functions for instruments and classes of instruments.
 Each set of methods is contained within a subpackage of this set.
 """
 
-from . import demeter, madrigal, nasa_cdaweb, sw
+from . import demeter, madrigal, nasa_cdaweb, sw, testing
