+++ pysat/instruments/pysat_testing.py
@@ -11,7 +11,7 @@ import numpy as np
 import pandas as pds
 
 import pysat
-from pysat.instruments import testing_methods as test
+from pysat.instruments.methods import testing as test
 
 # pysat required parameters
 platform = 'pysat'
